// langgraph/stockbroker/types.ts
const stockPurchaseSchema = z.object({
  ticker: z.string().describe("The ticker of the stock. Example: 'AAPL'"),
  quantity: z
    .number()
    .int()
    .positive()
    .describe("The quantity of stock to purchase."),
  maxPurchasePrice: z
    .number()
    .positive()
    .describe(
      "The max price at which to purchase the stock. Defaults to the current price."
    ),
});

export type StockPurchase = z.infer<typeof stockPurchaseSchema>;